pyproject.toml
[project]
name = "pyhk3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Gunther Klessinger", email = "gklessinger@gmx.de" }]
dependencies = [
  "absl-py>=2.1.0",
  "pyyaml>=6.0.2",
  "requests>=2.32.3",
  "rich>=13.9.4",
  "sh>=2.2.1",
  "structlog>=25.1.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.sdist]
include = ["src/pyhk3"]


[dependency-groups]
dev = ["pytest>=8.3.4"]


[project.scripts]
pyhk3 = "pyhk3.cli:main"


[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
# ruff is enough
reportSelfClsParameterName = false
pythonPlatform = "Linux"
typeCheckingMode = "off"
reportMissingImports = true
reportMissingTypeStubs = false
reportUndefinedVariable = true
reportGeneralTypeIssues = false
reportUnusedExpression = false
root = "src"
include = ["src", "tests"]
# pythonVersion = "3.8"
executionEnvironments = [{ root = "src" }, { root = "tests" }]
venvPath = "./"
venv = ".venv"

[tool.ruff]
# https://docs.astral.sh/ruff/rules/
line-length = 90
extend-select = ["Q"]
select = ["E", "F", "B"] # Enable flake8-bugbear (`B`) rules.
ignore = [
  "E501", # Never enforce `E501` (line length violations).
  "E741", # short var names
  "E731", # no lambda
  "B006", # mutables in signature
]

[tool.ruff.lint]
fixable = ["ALL"]
unfixable = [
  "B",    # Avoid trying to fix flake8-bugbear (`B`) violations.
  "F401", # Unused Import
  "F841", # variable assigned but not used
]


[tool.ruff.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "single"

[tool.ruff.format]
# Prefer single quotes over double quotes
quote-style = "single"
